Used clang static analyzer and additional warning levels to improve code.

+++ ctrl-lib/misc/CircleGenerator.cpp
@@ -44,7 +44,7 @@ namespace CalX {
 		return this->center;
 	}
 
-	int Circle::getRadius() {
+    int64_t Circle::getRadius() {
 		return this->radius;
 	}
